// src/app/models/general-information.dto.ts
export class GeneralInformationDto {
  HeaderText: string;
  NameSurname: string;
  ShortText: string;
  BirthDate: string;
  Address: string;
  EMail: string;
  Phone: string;
  AboutMe: string;
  GoogleMap: string;
  Facebook: string;
  Instagram: string;
  Twitter: string;
  Skype: string;
  SkypeConnectionType: string;
  Tumblr: string;
  Youtube: string;
  Linkedin: string;
}