// src/app/models/site-settings.dto.ts
export class SiteSettingsDto {
  SiteTitle: string;
  MetaKeys: string;
  Description: string;
  Theme: string;
  Hex: string;
  Css: string;
  AboutMe: number;
  Experiences: number;
  Skills: number;
  Projects: number;
  Educations: number;
  Message: number;
  MyReferences: number;
  Blog: number;
  Awards: number;
}